Getting bypassed (plugins) to work.  Some fixes for KDE2 building too.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@3108 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/loopruler.cpp
@@ -61,7 +61,7 @@ LoopRuler::~LoopRuler()
 
 void LoopRuler::scrollHoriz(int x)
 {
-    m_currentXOffset = -x / getHScaleFactor();
+    m_currentXOffset = int(double(-x)/getHScaleFactor());
 
     repaint();
 }
